@@ -5,29 +5,29 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.5</version>
+        <version>3.2.5</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.giova.service</groupId>
     <artifactId>moneystats</artifactId>
-    <version>2.1.1</version>
+    <version>2.2.0</version>
     <!-- <packaging>war</packaging> -->
     <packaging>jar</packaging>
     <name>MoneyStats-Service</name>
     <description>MoneyStats Service API</description>
     <properties>
-        <java.version>11</java.version>
+        <java.version>22</java.version>
     </properties>
     <dependencies>
         <dependency>
             <groupId>io.github.giovannilamarmora.utils</groupId>
             <artifactId>utils-code</artifactId>
-            <version>0.2.77</version>
+            <version>1.0.35</version>
         </dependency>
         <dependency>
             <groupId>org.apache.tomcat.embed</groupId>
             <artifactId>tomcat-embed-core</artifactId>
-            <version>9.0.68</version> <!-- Do Not Update -->
+            <version>10.1.26</version> <!-- Do Not Update -->
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
