@@ -2,11 +2,11 @@ package com.giova.service.moneystats.app.category;
 
 import com.giova.service.moneystats.app.category.dto.Category;
 import com.giova.service.moneystats.app.category.entity.CategoryEntity;
+import io.github.giovannilamarmora.utils.context.TraceUtils;
 import io.github.giovannilamarmora.utils.generic.Response;
 import io.github.giovannilamarmora.utils.interceptors.LogInterceptor;
 import io.github.giovannilamarmora.utils.interceptors.LogTimeTracker;
 import io.github.giovannilamarmora.utils.interceptors.Logged;
-import io.github.giovannilamarmora.utils.interceptors.correlationID.CorrelationIdUtils;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -29,8 +29,7 @@ public class CategoryService {
     String message = "List of category!";
 
     Response response =
-        new Response(
-            HttpStatus.OK.value(), message, CorrelationIdUtils.getCorrelationId(), categories);
+        new Response(HttpStatus.OK.value(), message, TraceUtils.getSpanID(), categories);
 
     return ResponseEntity.ok(response);
   }
