@@ -4,10 +4,10 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.giova.service.moneystats.app.stats.dto.Stats;
 import com.giova.service.moneystats.crypto.asset.dto.Asset;
 import io.github.giovannilamarmora.utils.generic.GenericDTO;
+import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
-import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
