@@ -3,9 +3,9 @@ package com.giova.service.moneystats.crypto.coinGecko;
 import com.giova.service.moneystats.crypto.coinGecko.entity.MarketDataEntity;
 import io.github.giovannilamarmora.utils.interceptors.LogInterceptor;
 import io.github.giovannilamarmora.utils.interceptors.LogTimeTracker;
+import jakarta.transaction.Transactional;
 import java.util.List;
 import java.util.Objects;
-import javax.transaction.Transactional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
